Some worker refactor (partial DM-548, partial DM-1145)
Removed currently disused worker code, mostly relating to the
previous XRootD file-based query mechanism.

Gave xrdname class a new home in module xrdsvc (the only module
that uses it, and its previous home xrdfs was going away).

Given other deletions, wcontrol::Service became a trivial
forieman-specific pass-through, so removed this too.

The scheduler for the foreman is now constructed externally,
and passed to the foreman at construction time (this breaks
the previous include-cycle wcontrol <-> wsched).

Added missing virtual dtor to class PacketBuffer::Source.

ATTIC: modules xrdc, xrdfs
ATTIC: class wbase::TaskAccepter
ATTIC: classes wcontrol::RequestTaker, wcontrol::ResultTracker, wcontrol::Service
ATTIC: class ccontrol::thread
ATTIC: class wdb::QueryRunner
