+++ core/modules/czar/SConscript
@@ -29,8 +29,7 @@ depends = ['ccontrol/dispatcher.h',
            'qproc/ChunkSpec.h',
            'rproc/TableMerger.h',
            'rproc/mergeTypes.h',
-           'util/Substitution.h',
-           'xrdc/xrdfile.h']
+           'util/Substitution.h']
 depends = map(lambda f: env.File("../"+f), depends)
 
 pythonInc = distutils.sysconfig.get_python_inc()
