+++ core/modules/wbase/Task.h
@@ -75,17 +75,13 @@ public:
     };
 
     explicit Task() {}
-    explicit Task(wbase::ScriptMeta const& s, std::string const& user_=defaultUser);
-    explicit Task(TaskMsgPtr t, std::string const& user_=defaultUser);
     explicit Task(TaskMsgPtr t, boost::shared_ptr<wbase::SendChannel> sc);
 
     TaskMsgPtr msg; ///< Protobufs Task spec
     boost::shared_ptr<wbase::SendChannel> sendChannel; ///< For result reporting
     std::string hash; ///< hash of TaskMsg
     std::string dbName; ///< dominant db
-    std::string resultPath; ///< Result path for non-SendChannel operation
     std::string user; ///< Incoming username
-    bool needsCreate; ///< Deprecated.
     time_t entryTime; ///< Timestamp for task admission
     char timestr[100]; ///< ::ctime_r(&t.entryTime, timestr)
     // Note that manpage spec of "26 bytes"  is insufficient
