+++ core/modules/wcontrol/Foreman.cc
@@ -43,11 +43,10 @@
 #include "proto/worker.pb.h"
 #include "wbase/Base.h"
 #include "wbase/MsgProcessor.h"
+#include "wbase/SendChannel.h"
 #include "wconfig/Config.h"
 #include "wdb/ChunkResource.h"
 #include "wdb/QueryAction.h"
-#include "wdb/QueryRunner.h"
-#include "wsched/FifoScheduler.h"
 
 ////////////////////////////////////////////////////////////////////////
 // anonymous helpers
@@ -84,7 +83,6 @@ public:
     virtual ~ForemanImpl();
 
     bool squashByHash(std::string const& hash);
-    bool accept(boost::shared_ptr<proto::TaskMsg> msg);
     void newTaskAction(wbase::Task::Ptr task);
 
     virtual boost::shared_ptr<wbase::MsgProcessor> getProcessor();
@@ -99,8 +97,6 @@ public:
         std::string const& getHash() const { return _task->hash; }
 
     private:
-        void _runWithDump();
-        void _runProtocol2();
         RunnerMgr& _rm;
         wbase::Task::Ptr _task;
         bool _isPoisoned;
@@ -112,7 +108,6 @@ public:
     public:
         RunnerMgr(ForemanImpl& f);
         void registerRunner(Runner* r, wbase::Task::Ptr t);
-        boost::shared_ptr<wdb::QueryRunner> newQueryRunner(wbase::Task::Ptr t);
         boost::shared_ptr<wdb::QueryAction> newQueryAction(wbase::Task::Ptr t);
         void reportComplete(wbase::Task::Ptr t);
         void reportStart(wbase::Task::Ptr t);
@@ -149,11 +144,8 @@ private:
 ////////////////////////////////////////////////////////////////////////
 Foreman::Ptr
 newForeman(Foreman::Scheduler::Ptr sched) {
-    if(!sched) {
-        sched = boost::make_shared<wsched::FifoScheduler>();
-    }
     ForemanImpl::Ptr fmi = boost::make_shared<ForemanImpl>(sched);
-    return fmi;;
+    return fmi;
 }
 
 ////////////////////////////////////////////////////////////////////////
@@ -189,13 +181,6 @@ ForemanImpl::RunnerMgr::newQueryAction(wbase::Task::Ptr t) {
     wdb::QueryActionArg a(_f._log, t, _f._chunkResourceMgr);
     boost::shared_ptr<wdb::QueryAction> qa = boost::make_shared<wdb::QueryAction>(a);
     return qa;
-}
-
-boost::shared_ptr<wdb::QueryRunner>
-ForemanImpl::RunnerMgr::newQueryRunner(wbase::Task::Ptr t) {
-    wdb::QueryRunnerArg a(_f._log, t);
-    boost::shared_ptr<wdb::QueryRunner> qr = boost::make_shared<wdb::QueryRunner>(a);
-    return qr;
 }
 
 void
@@ -259,9 +244,6 @@ ForemanImpl::RunnerMgr::getLog() {
 class matchHash {
 public:
     matchHash(std::string const& hash_) : hash(hash_) {}
-    inline bool operator()(wdb::QueryRunnerArg const& a) {
-        return a.task->hash == hash;
-    }
     inline bool operator()(ForemanImpl::Runner const* r) {
         return r->getHash() == hash;
     }
@@ -305,26 +287,16 @@ void ForemanImpl::Runner::poison() {
     _isPoisoned = true;
 }
 
-void ForemanImpl::Runner::_runWithDump() {
-    boost::shared_ptr<wdb::QueryRunner> qr;
-    qr = _rm.newQueryRunner(_task);
-    qr->actOnce();
-}
-
-void ForemanImpl::Runner::_runProtocol2() {
-    _action = _rm.newQueryAction(_task);
-    (*_action)();
-}
-
 void ForemanImpl::Runner::operator()() {
     _rm.registerRunner(this, _task);
     while(!_isPoisoned) {
         LOGF(_log, LOG_LVL_INFO, "Runner running %1%" % *_task);
         proto::TaskMsg const& msg = *_task->msg;
-        if(msg.has_protocol() && msg.protocol() == 2) {
-            _runProtocol2();
+        if(!msg.has_protocol() || msg.protocol() < 2) {
+            _task->sendChannel->sendError("Unsupported wire protocol", 1);
         } else {
-            _runWithDump();
+            _action = _rm.newQueryAction(_task);
+            (*_action)();
         }
         if(_isPoisoned) break;
         // Request new work from the manager
@@ -392,21 +364,7 @@ bool ForemanImpl::squashByHash(std::string const& hash) {
     boost::lock_guard<boost::mutex> m(_mutex);
     bool success = _scheduler->removeByHash(hash);
     success = success || _rManager->squashByHash(hash);
-    if(success) {
-        // Notify the tracker in case someone is waiting.
-        ResultError r(-2, "Squashed by request");
-        wdb::QueryRunner::getTracker().notify(hash, r);
-        // Remove squash notification to prevent future poisioning.
-        wdb::QueryRunner::getTracker().clearNews(hash);
-    }
     return success;
-}
-
-bool
-ForemanImpl::accept(boost::shared_ptr<proto::TaskMsg> msg) {
-    wbase::Task::Ptr t = boost::make_shared<wbase::Task>(msg);
-    newTaskAction(t);
-    return true;
 }
 
 void ForemanImpl::newTaskAction(wbase::Task::Ptr task) {
