+++ core/modules/wcontrol/Foreman.h
@@ -69,7 +69,7 @@ namespace wcontrol {
 
 /// Foreman is a pooling thread manager that is pluggable with different
 /// scheduling objects
-class Foreman : public wbase::TaskAcceptor {
+class Foreman {
 public:
     typedef boost::shared_ptr<Foreman> Ptr;
 
@@ -102,8 +102,6 @@ public:
 
     virtual bool squashByHash(std::string const& hash) { return false; }
 
-    /// Accept a new task for execution
-    virtual bool accept(boost::shared_ptr<proto::TaskMsg> msg) { return false; }
     virtual boost::shared_ptr<wbase::MsgProcessor> getProcessor() = 0;
     virtual ~Foreman() {}
