+++ core/modules/xrdsvc/SsiService.h
@@ -36,7 +36,7 @@ class XrdSsiLogger;
 namespace lsst {
 namespace qserv {
 namespace wcontrol {
-  class Service;
+  class Foreman;
 }
 namespace wpublish {
   class ChunkInventory;
@@ -67,7 +67,7 @@ private:
     bool _setupScratchDb();
 
     boost::shared_ptr<wpublish::ChunkInventory> _chunkInventory;
-    boost::shared_ptr<wcontrol::Service> _service;
+    boost::shared_ptr<wcontrol::Foreman> _foreman;
 
 }; // class SsiService
